:recycle: refactor: store entry to entry table && fetch rss api

+++ greader_api/type_entry.go
@@ -33,7 +33,7 @@ type Entry struct {
 	Title  string `json:"title"`
 	Author string `json:"author"`
 
-	PublishedTimestamp int    `json:"published"`
+	PublishedTimestamp int64  `json:"published"`
 	CrawledTimestamp   string `json:"crawlTimeMsec"`
 	TimestampUsec      string `json:"timestampUsec"`
